requirements.txt
flask==3.0.3
requests==2.32.3
opencv-python==4.10.0.84
python-dotenv==1.0.1
